services/kyc-verifier/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
deepface==0.0.93
easyocr==1.7.2
python-multipart==0.0.12
Pillow>=10.0
tf-keras
